openvibe-kernel :  + added distinction between boxes that were suspended because of a return false or flagged as crashed because of successive crashes  * corrected bug on visualisation tree on x64 arch (an was enum casted to unsigned long in sscanf)
git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@1529 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe-kernel-omk/trunc/src/openvibe-kernel/kernel/visualisation/ovkCVisualisationTree.cpp
@@ -399,7 +399,7 @@ boolean CVisualisationTree::reloadTree()
 	gtk_tree_store_set(m_pTreeStore, &l_oIter,
 		EVisualisationTreeColumn_StringName, "Unaffected display plugins",
 		EVisualisationTreeColumn_StringStockIcon, m_pTreeViewCB->getTreeWidgetIcon(EVisualisationTreeNode_Unaffected),
-		EVisualisationTreeColumn_ULongNodeType, EVisualisationTreeNode_Unaffected,
+		EVisualisationTreeColumn_ULongNodeType, (unsigned long)EVisualisationTreeNode_Unaffected,
 		EVisualisationTreeColumn_StringIdentifier, (const char*)OV_UndefinedIdentifier.toString(),
 		-1);
 
@@ -848,7 +848,7 @@ boolean CVisualisationTree::loadVisualisationWidget(IVisualisationWidget* pVisua
 	gtk_tree_store_set(m_pTreeStore, &l_oIter,
 		EVisualisationTreeColumn_StringName, (const char*)pVisualisationWidget->getName(),
 		EVisualisationTreeColumn_StringStockIcon, (const char*)l_pStockIconString,
-		EVisualisationTreeColumn_ULongNodeType, l_oChildType,
+		EVisualisationTreeColumn_ULongNodeType, (unsigned long)l_oChildType,
 		EVisualisationTreeColumn_StringIdentifier, (const char*)pVisualisationWidget->getIdentifier().toString(),
 		EVisualisationTreeColumn_PointerWidget, l_pWidget,
 		-1);
